// src/services/openapi/models/UserOutput.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserOutput = {
    id: string;
    email: string;
    firstName: string;
    lastName: string;
    login: string;
    last_connexion_date?: string | null;
    roles: Array<string>;
    created_at: string;



};